Increases security rocket launcher explosion damage from 30 to 50 because I got shot by it once and was like "wait that's it" (#4905)
## About The Pull Request

Increases security rocket launcher explosion damage from 30 to 50
because I got shot by it once and was like "wait that's it"

## Why It's Good For The Game

The weapon is super weak and no one uses it. Hopefully a numbers change
gets people to use it.

## Proof Of Testing

If it compiles, it werks.

</details>

## Changelog

:cl: BurgerBB
balance: Increases security rocket launcher explosion damage from 30 to
50
/:cl: