Fixes the Moonstation Gamer Lair causing map errors. (#4934)
## About The Pull Request

Revamps the Moon Station gamer lair in maint near cargo to prevent map
errors.

## Why It's Good For The Game

Bitrunning only supports 3 pods now, and there is an automatic thing to
remove extra pods, which sometimes causes map errors because the
bitrunning pod in the gaming den has a marker that makes it start
broken.

## Proof Of Testing

If it compiles, it werks.

## Changelog

:cl: BurgerBB
map: Revamps the Moon Station gamer lair in maint near cargo to prevent
map errors.
/:cl: