Fixes penis examine (#4955)
## About The Pull Request

Penises now show the description they are supposed to.
## Why It's Good For The Game

For a while now examining penises has only showed the size, not the type
or if there was a slit/sheath. This fixes that. Fixes are good.
## Proof Of Testing

Works locally. Tested both with and without a slit.
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl: Midnight Starchaser
fix: fixed penis examine text
/:cl:

Co-authored-by: Alexis <catmc8565@gmail.com>