[SEMI-MODULAR] Oversized SAD Functionality (#4947)
## About The Pull Request
Adds a new variable on `/obj/machinery` called `ignore_size` it does as
implied, ignores the size of a mob if the variable is set to true on a
machine. So far, I only implemented this on the SAD so that oversized
quirk users can finally use it, but it can be modularly added to pretty
much any machine that opens/closes.
## Why It's Good For The Game
Oversized quirk users can finally use the SAD instead of being screwed
and having to rely on getting a NIF (and being completely screwed in a
few odd edge cases).
## Proof Of Testing
Parsed, compiled, and ran without any runtimes and worked as expected.
## Changelog
:cl:
qol: Oversized quirk users can now use SADs.
/:cl: