The 'Request OPFOR' verb now requires confirmation (#4978)
## About The Pull Request
Does as the title says. When you press the button, a pop-up now asks if
you are sure you want to press it.
## Why It's Good For The Game
I've watched a handful of staff members accidentally hit it in my time
here while trying to play web sounds, enough that I feel it warrants
this. Especially given that it pings every player with be antagonist
turned on.
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>


https://github.com/user-attachments/assets/f59a26ec-5507-4cbb-b9f1-6c1300b1d10e

</details>

## Changelog
:cl:
admin: The Request OPFOR verb now asks for confirmation before pinging
every antagonist enabled player for OPFORs.
/:cl: