Adds an ERP items global list (#5036)
## About The Pull Request
It's a (to my knowledge) complete list of all the items (structures,
machines, etc. excluded). I need it for some of my other PRs I have
planned (or already up), but I wanted to atomize.
## Why It's Good For The Game
It's nothing player facing, but it makes things easier when trying to do
things like blacklists or refactoring certain things.
## Proof Of Testing
It's literally just a list, it works.
## Changelog
Nothing player facing.