Loadout Painting Printer (#4649)
## About The Pull Request
Adds a new modular computer type to the loadout screen: a mini painting
printer with a pre-installed program that prints the selected art, with
a painting frame to match. It can only print a single time, then
self-destructs.
## Why It's Good For The Game
Allows spotlight on the... furry art, in our painting database, without
having to rely on the half-present curator role.
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
add: adds a single-user painting printer to the loadout menu and the fun
vendor
/:cl:

---------

Co-authored-by: The Sharkening <95130227+StrangeWeirdKitten@users.noreply.github.com>
Co-authored-by: Alexis <catmc8565@gmail.com>