Adds a Quirk: "Loose Limbs" That Causes Your Limbs To Fall Off On Death! (#5030)
## About The Pull Request
Full disclosure: This is a PR commissioned by CodeRedAlert on the
discord for 20 USD.

Adds a quirk which causes your limbs (arms and legs) to fall off on
death. This is a feature from the "Cursed" quirk that the commissioner
wanted to see isolated. We agreed that -2 seems like a fair point value
to give it, but this can be open to change, either making it worth more
or less points if the maintainers agree on it.
Tested locally, as far as I know, there are no bugs with it, and it
should not inherit any other of the Cursed quirk.
## Why It's Good For The Game
Adds extra flavor to players who want characters with weak limbs, such
as a piñata character, or a doll.

## Proof Of Testing
<img width="158" height="161" alt="image"
src="https://github.com/user-attachments/assets/f768ffe5-acd0-4861-99ed-d2dab83173fa"
/>

Those limbs sure do come off! Tested locally in a variety of
circumstances to ensure it functions as expected.

## Changelog
:cl:
add: New quirk- Loose Limbs!
/:cl: