Adds two new goody packs (crew monitor/soap) (#5066)
## About The Pull Request

Adds two new goody packs to be ordered in cargo. Crew monitor and bar of
soap.

## Why It's Good For The Game

Sometimes you want to purchase this without having to order an entire
nanomed refill. Soap is available for free as a loadout item, so this
makes it a bit more accessible.

## Changelog

:cl: LT3
add: Handheld crew monitor is now orderable via cargo
add: Soap is now orderable via cargo
/:cl: