Disables borers because I died to it once. (#4985)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

I made the weight 0 and made the minimum player count 999. I'm assuming
that'll stop it from happening.

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

Borers are an unmaintained antagonist, they need a ton of work to bring
them into an acceptable or balanced state. Something something "Improve,
don't remove", but if leaving these enabled as-is causes a lot of
headache and disruption considering how outdated they are.

As it stands now, they can give themselves free respawns by hiding eggs,
act as free stimulant machines, dump infinite death chems into people at
a moments notice, develop immunity to their primary weakness, and gain
stupid amounts of health.

If a borer is bored or hates you, enjoy getting brain damage and organ
failure while they pump an endless slurry of chemicals in your body
until they get yanked out by an overworked medbay. It's not fun being
taken out of a round for 30 minutes because something click-dragged
itself into you while you were stunned by a sec officer or slipped on a
tile of water. It's even less fun when they can do it again the moment
you leave medbay because there was an egg stored somewhere.

## Proof Of Testing

This is one of those things that can't really be tested locally, I don't
think, but it's just a numbers change and I read the datums so I knew to
change minimum player counts instead of weights.

</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
del: Borers are (probably) disabled.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->