Disables all Heretic Robes to be swapped to non-functional mode (neck slot) (#5053)
## About The Pull Request

Could fix: https://github.com/Bubberstation/Bubberstation/issues/5050

This PR adds all heretic robes to the `only_functional` blacklist that
makes the robe unable to be switched to Non-Functional mode. This is a
soft hack-fix but ultimately should be implemented anyways. There is an
issue preventing specifically heretic robes from being stripped if
equipped on the neck slot; thus making the wearer put the hood on
whenever for the focus.

This PR prevents all heretic robes to be used on the neck slot and must
be worn on the suit slot.

Additionally, I've added a message sent to your chat if you attempt to
swap a blacklisted piece of suit into non-functional mode.

## Why It's Good For The Game

This keeps the forced functional blacklist updated with items that
should remain functional due to its mechanical use.

## Proof Of Testing

<details>
<summary>Screenshots/Videos</summary>


https://github.com/user-attachments/assets/7b6e0d9b-69b9-41c9-943f-5ada9c634b68

</details>

## Changelog

:cl:
fix: Fixed heretic robes functioning on the neck slot.
code: Added a message to the chat when a suit can't swap to
non-functional mode.
/:cl:

---------

Co-authored-by: LT3 <83487515+lessthnthree@users.noreply.github.com>