turns dom aura + well trained into neutral quirks (#5074)
## About The Pull Request
No longer shall they give points
## Why It's Good For The Game

These quirks have been a longterm cause for confusion, primarily about
their usability within combat. They are quirks that give positive and
negative values each, so in the past it has been ruled that this makes
them okay to use in combat

However, they are also labeled as `erp_quirk = TRUE`, every other bit of
ERP content in the game is barred from being used to gain a mechanical
advantage. I.E. The sex whip. This being an exception does not help with
the confusion, and ideally we want as few exceptions as possible.
## Proof Of Testing
<details>
<summary>Screenshots/Videos</summary>

</details>

## Changelog
:cl:
balance: Well Trained and Dominant Aura are now neutral quirks.
/:cl: